Form group: desktop alignment.
