+++ packages/form/src/fields/Group/index.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import classNames from 'classnames';
 
-import type { FieldProps } from '../types';
-import Field from './Field';
-import FieldRenderer from '../renderers/FieldRenderer';
-import useShouldBeVisible from '../hooks/useShouldBeVisible';
+import Field from '../Field';
+import FieldRenderer from '../../renderers/FieldRenderer';
+import useShouldBeVisible from '../../hooks/useShouldBeVisible';
+import type { FieldProps } from '../../types';
+
+import './styles.scss';
 
 const Group: React.FC<FieldProps> = ({
+	columns,
 	subFields,
 	label,
 	name: groupName,
@@ -16,23 +19,29 @@ const Group: React.FC<FieldProps> = ({
 	formControlProps,
 }) => {
 	const visible = useShouldBeVisible(conditions, groupName);
+
+	console.log({ columns });
+
 	if (subFields.length && visible) {
 		// className for the group
 		const className = formControlProps?.className;
-		const groupClassName = classNames('field-group-items', className);
-		// className for the group wrapper
-		const wrapperClassName = classNames('field-group-wrapper', {
+		const groupClassName = classNames('ee-field-group__items ee-row', className);
+		const itemClassName = classNames('ee-field-group__item', columns && `ee-col--${columns}`);
+
+		const wrapperClassName = classNames('ee-field-group__wrapper', {
 			[`${className}-wrapper`]: className,
 		});
+
 		return (
 			<div className={wrapperClassName}>
-				{label && <h5 className='field-group-label'>{label}</h5>}
+				{label && <h5 className='ee-field-group__label'>{label}</h5>}
 				{before}
 				<div className={groupClassName}>
 					{subFields.map(({ name: fieldname, fieldType, ...props }, i) => {
 						const name = `${groupName}.${fieldname}`;
+
 						return (
-							<div className='field-group-item' key={name + i}>
+							<div className={itemClassName} key={name + i}>
 								<Field component={FieldRenderer} {...props} fieldType={fieldType} name={name} />
 							</div>
 						);
