+++ packages/form/src/stories/TestForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { formConfig } from './config';
-import EspressoForm from './EspressoForm';
+import EspressoForm from '../EspressoForm';
 
 const subscription = {};
